Commit type: style / fix
Home Section responsive issue for devices under 1200px breakpoint

Making the home section responsive for under 1200px wide devices. Responsive layout is mobile-first.  Order of The HTML code of this section was written with the desktop design in mind; but in mobile mode (with width of 992px and lower) the order of elements changes; Picture is on top and the rest under. But in desktop, picture is on the right side and the rest on left side. Font size of the title changes as the width of the screen decreases. Row gap between left elements reduces as you go down from 2xl (1400px) breakpoint. the products shown underneath the person's picture, appears in 1200px and above breakpoints. Chat bubble in shown in 768px breakpoints and beyond.