Favorite-Category Section implementation and making it responsive
Commit type: style / fix

The second section in the home page, customer favorites, is about different categories; 4 in total. Using gris system, 4 columns were made. Each category is a card; the cards are components and can be used later on as well. The appearance of the cards don't change much from breakpoints to another; except below 768px, in which the flex direction of the card, changes from row to column.