Favorite-Category Section implementation and making it responsive
Commit type: style

The section for different dishes with the slider is complete. Using Swiper js, I was able to make the slider responsively; It fully adapts to the window's width and changes appearance accordingly. I utilized a different form of navigation buttons for this slider. Swiper provides navigation buttons, but this project has a set of its own.