Final Commit Of The Project
Commit type: final commit

All styles are gathered in the main.css file. It is minified and compressed for faster loading time and better overall performance of the website. All images used in the project are optimized and minimized to reduce loading wait time. There is a lot of unused svg files that were deleted from the project directory. unnecessary files are deleted.